requirements.txt
discord.py
ptyhon--dotenv
requests